requirements.txt
psycopg2-binary==2.9.9
streamlit==1.28.1
streamlit-option-menu==0.3.6
pandas==2.1.3
pillow==10.1.0
pdf2image==1.16.3
requests==2.31.0
python-dotenv==1.0.0 